Further splitting of inner classes
