+++ src/Applications/UUPDownload/DownloadRequestOptions.cs
@@ -117,90 +117,4 @@ namespace UUPDownload
             get; set;
         }
     }
-
-    [Verb("replay-download", isDefault: false, HelpText = "Replay a download from zero using a *.uupmcreplay file.")]
-    internal class DownloadReplayOptions
-    {
-        [Option('r', "replay-metadata", HelpText = @"The path to a *.uupmcreplay file to replay an older update and resume the download process. Example: D:\20236.1005.uupmcreplay", Required = true)]
-        public string ReplayMetadata
-        {
-            get; set;
-        }
-
-        [Option('t', "machine-type", HelpText = "The architecture to report to the Windows Update servers. Example: amd64", Required = true)]
-        public MachineType MachineType
-        {
-            get; set;
-        }
-
-        [Option('o', "output-folder", HelpText = "The folder to use for downloading the update files.", Required = false, Default = ".")]
-        public string OutputFolder
-        {
-            get; set;
-        }
-
-        [Option('e', "edition", HelpText = "The edition to get. Must be used with the language parameter. Omit either of these to download everything. Example: Professional", Required = false, Default = "")]
-        public string Edition
-        {
-            get; set;
-        }
-
-        [Option('l', "language", HelpText = "The language to get. Must be used with the edition parameter. Omit either of these to download everything. Example: en-US", Required = false, Default = "")]
-        public string Language
-        {
-            get; set;
-        }
-
-        [Option("fixup", HelpText = @"Applies a fixup to files in output folder. Example: Appx", Required = false)]
-        public Fixup? Fixup
-        {
-            get; set;
-        }
-
-        [Option("appxroot", HelpText = @"The folder containing the appx files for use with the Appx fixup", Required = false)]
-        public string AppxRoot
-        {
-            get; set;
-        }
-
-        [Option("cabsroot", HelpText = @"The folder containing the cab files for use with the Appx fixup", Required = false)]
-        public string CabsRoot
-        {
-            get; set;
-        }
-    }
-
-    [Verb("get-builds", isDefault: false, HelpText = "Get builds in all rings matching the request type")]
-    internal class GetBuildsOptions
-    {
-        [Option('s', "reporting-sku", HelpText = "The sku to report to the Windows Update servers. Example: Professional", Required = true)]
-        public OSSkuId ReportingSku
-        {
-            get; set;
-        }
-
-        [Option('t', "machine-type", HelpText = "The architecture to report to the Windows Update servers. Example: amd64", Required = true)]
-        public MachineType MachineType
-        {
-            get; set;
-        }
-
-        [Option('m', "mail", HelpText = "Email for the Windows Insider account to use to generate authorization tokens (Optional)", Required = false, Default = "")]
-        public string Mail
-        {
-            get; set;
-        }
-
-        [Option('p', "password", HelpText = "Password for the Windows Insider account to use to generate authorization tokens (If 2FA, must be generated app password) (Optional)", Required = false, Default = "")]
-        public string Password
-        {
-            get; set;
-        }
-
-        [Option("preview-targeting-attribute", HelpText = "The name of the set of targeting attributes to use. (Optional, Preview)", Required = false, Default = "")]
-        public string TargetingAttribute
-        {
-            get; set;
-        }
-    }
 }
